+++ test/specs/modules/Dropdown/DropdownHeader-test.js
@@ -4,4 +4,5 @@ import * as common from 'test/specs/commonTests'
 describe('DropdownHeader', () => {
   common.isConformant(DropdownHeader)
   common.rendersChildren(DropdownHeader)
+  common.implementsIconProp(DropdownHeader)
 })